master/requirements.txt
aiohttp==3.7.3
pyaml==20.4.0
pydantic==1.7.2
pytest==6.1.2
PyYAML==5.3.1
toml==0.10.2
trafaret==2.1.0
typing-extensions==3.7.4.3
loguru==0.5.3
async-timeout==3.0.1